adds mark as read notification

+++ db/migrate/20200416151013_create_notifications.rb
@@ -6,8 +6,8 @@ class CreateNotifications < ActiveRecord::Migration[6.0]
       t.datetime :read_at
       t.string :action
       t.string :notifiable_type
-      t.integer :notifiable_id
-
+      t.bigint :notifiable_id
+  
       t.timestamps
     end
   end
